2014-01-10 Roman - !!!!:SYS,UI.{QTStarter,QTCfg,Vision}: Adapted for build with Qt5, exclude Phonon. - SYS: Warning messages drop for no "gdb" present.     User's API functions gmtime(), timegm(), strftimegm(), strptimegm() for time into GMT(UTC) processing allow. - SYS,DB,TVariant,DAQ.JavaLikeCalc,Special.FLibSYS: Direct access, into backend, to user space object TArrayObj is set. - TVariant,DAQ.{JavaLikeCalc,Siemens}: Integer type expanded next to int64_t. - FIX:DB.{SQLite,MySQL,FireBird,PostgreSQL}: MTable::fieldDel() is fixed for no present-translated fields process. - DAQ.JavaLikeCalc: Direct access to object's properties set for some performance rise. - FIX:DAQ.OPC_UA: LifeTime for session processing into client's part is fixed and SC_RENEW support is added. - Special.FLibSYS: Object IO "write" implementation finished.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2034 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/Siemens/siemens.cpp
@@ -1220,7 +1220,7 @@ char TMdContr::getValB( SValData ival, ResString &err )
     return EVAL_BOOL;
 }
 
-int TMdContr::getValI( SValData ival, ResString &err )
+int64_t TMdContr::getValI( SValData ival, ResString &err )
 {
     int iv_sz = valSize( IO::Integer, ival.sz );
     for(unsigned i_b = 0; i_b < acqBlks.size(); i_b++)
@@ -1239,7 +1239,7 @@ int TMdContr::getValI( SValData ival, ResString &err )
 	}
     if(err.getVal().empty()) err.setVal( _("11:Value is not gathered.") );
 
-    return EVAL_INT;
+    return (int64_t)EVAL_INT;
 }
 
 double TMdContr::getValR( SValData ival, ResString &err )
@@ -1306,7 +1306,7 @@ void TMdContr::setValB( bool ivl, SValData ival, ResString &err )
     catch(TError cerr) { if(err.getVal().empty()) err.setVal(cerr.mess); }
 }
 
-void TMdContr::setValI( int ivl, SValData ival, ResString &err )
+void TMdContr::setValI( int64_t ivl, SValData ival, ResString &err )
 {
     int val = getValI(ival, err);
     if(val == EVAL_INT || val == ivl) return;
